Progressive download in Phonon MMF backend: integrated with player
This commit integrates the Download class with the media playback
classes in the backend, to implement Progressive Download.

Note that this PDL implementation has one drawback: when video
playback is paused due to shortage of data (i.e. due to the download
being temporarily stalled), the display goes black.  This is because,
when the end of the currently-downloaded data is reached, the
playback session is closed.  When more data becomes available, the
clip is re-opened, a seek is done to reach the previous playback
position, and playback is re-started.  Closing the playback session
closes the video stack's connection to the display, thereby causing
the video widget to go black while more data is buffered.

This is a consequence of the level in the native video stack at which
the Phonon integration is done: managing a network stall without
requiring the playback session to be closed would require integration
below the MMF client API, specifically at the MMF controller level.

Task-number: QTBUG-10769
Reviewed-by: Derick Hawcroft